漏斗图fixed
